Questionnaire Project/BBP_GENERATION/requirements.txt
# python-dotenv
# langchain
# langchain-openai
# langchain-community
# openai
# docx2txt
# python-docx
# chromadb

streamlit
python-docx
langchain
langchain-community
langchain-core
langchainhub
openai
chromadb
python-dotenv
tiktoken
docx2txt
pypdf
pydantic
httpx
watchdog
